refactor: split engine/coordination/dispatchers.py into per-dispatcher modules
